pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scriptoza"
version = "0.1.0"
description = "Modular utility scripts including VBC - Video Batch Compression with AV1/NVENC support"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.5",
    "pyexiftool>=0.5.6",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "textual>=0.50.0",
    "typer>=0.20.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.12.0",
]
docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "pymdown-extensions>=10.7",
]

[tool.hatch.build.targets.wheel]
packages = ["vbc"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
markers = [
    "slow: marks tests as slow (integration tests with real files)",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["vbc"]
omit = [
    "tests/*",
    "vbc/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
